+++ src/main_window.hpp
@@ -6,14 +6,19 @@
 #include <memory>
 
 class QCloseEvent;
+class QComboBox;
 class QLabel;
+class QLineEdit;
 class QResizeEvent;
+class QStackedWidget;
 class QSystemTrayIcon;
+class QTabBar;
 class QTimer;
 class QUrl;
 class QWebEngineNotification;
 class QWebEngineProfile;
 class QWebEngineView;
+class QWidget;
 
 namespace plainwire {
 
@@ -34,6 +39,7 @@ protected:
 
 private:
     void createProfile();
+    void createTabs();
     void createWebView();
     void createActions();
     void createTray();
@@ -44,12 +50,28 @@ private:
     void positionNotice();
     void setZoomFactor(double factor, bool announce = true);
 
+    QWidget *createServersPanel();
+    void refreshServerPanel();
+    void addServerFromUi();
+    void removeServerFromUi();
+    void applyServerFromUi();
+    void resetServerFromUi();
+    void openServersTab();
+    void switchToChat();
+    void applyServerScripts();
+    void reloadForServer();
+
     QWebEngineProfile *profile_ = nullptr;
     QWebEngineView *view_ = nullptr;
     PlainwirePage *page_ = nullptr;
     QSystemTrayIcon *tray_ = nullptr;
     QLabel *noticeLabel_ = nullptr;
     QTimer *noticeTimer_ = nullptr;
+    QTabBar *tabBar_ = nullptr;
+    QStackedWidget *stack_ = nullptr;
+    QWidget *serversPanel_ = nullptr;
+    QComboBox *serverCombo_ = nullptr;
+    QLineEdit *customHostEdit_ = nullptr;
     std::unique_ptr<QWebEngineNotification> notification_;
     bool quitting_ = false;
     bool recoveringRenderer_ = false;